refactor: Reduce Parallelism in CI
Our setup for this repository was to have lots of jobs run in parallel,
but we found the whole process to be a bit unstable. We wanted to see
if consuming less resources on every PR would help our situation.

Various changes have been made to enable this:
* Combine the parallel runs of all the different test categories into
  a single job and run each category in sequence.
* Remove the benchmark job from the all PR workflow and schedule it to
  run nightly.
* Remove the `cargo build` from the all PR workflow. The build will
  occur in the job that runs all the tests.
* Move ARM build and tests to nightly schedule rather than running for
  PRs.

Without caching, it takes ~35 minutes to build and run all the test
categories sequentially. The E2E tests take longer than this to run, so
this shouldn't be an issue.