refactor: Files Get Tests to Use Unique Data
Changes all the `files get` tests to use unique test data. This is
managed using the `assert_fs` crate, which will create temporary files
and directories, then automatically clean them up.

This was done because most of the tests were sharing the 'testdata'
directory, and a lot of the tests would delete this directory and
restore it again. The problem was, if one of the tests failed, this
would produce a cascade of failures, as other tests were expecting that
test data to exist.

Some of the functions in the test library were updated to accommodate
this, to allow the test to pass in directories they would have been
in control of. Some functions were removed because they weren't needed
any more.

These tests have been written to prioritise readability over reduction
of duplication. Therefore, the code that sets up the tests are often
duplicated. Obviously there is a maintenance cost here, but my
preference is for making the tests easier to understand without having
to jump into library functions to see what they're doing.

Each case was documented using the `given, when, then` style. This was
mainly just to help me understand what the tests were doing. There is
also obviously a maintenance cost with this, as if a test changes, the
documentation then also needs to be updated. The original comments were
kept as a synopsis.