fix: use proper dst_section_key when resend in case of AntiEntropyRedirect
When received AntiEntropyRedirect and need to re-send the bounced msg,
if selected elder is one of our local knowledge, use local knowledge as
dst_section_key, otherwise using genesis_key.
This is to ensure the AE flow for the resend message can be triggered
properly to updated our network knowledge.