feat: upgrade to 0.34.1 of sn_api
All references to SafeUrl have been renamed to Url, along with
SafeContentType and SafeDataType, which also lose the Safe prefix.

The CLI was still using u64 to represent versions of containers, but
this was replaced with VersionHash. There's some placeholder code at
the moment for dealing with the first 0 based version.

The `seq` command was completely removed, as the network no longer has
this concept any more.