fix: Redirect service messages from adults to elders
Previously, if an adult (non-elder) node received a `ServiceMsg` we
would silently ignore it. This commit changes the behaviour to instead
return an anti-entropy redirect to the client, redirecting them to the
current section's elders.

This might not be the desired long-term solution, since we don't really
want adults to be on the receiving end of messages from clients.