fix: Setting Sub Names Ignores Url Labels
New URLs for NrsMapContainer content were longer than 63 characters and
were failing validation when the URL was being parsed while fetching
NrsMapContainer content in `sn_api`.

The `set_sub_names` function is only being called in this single place
in the API, so the impact of this change should be quite low. It doesn't
break the unit test for the function.