chore: Upgrade safe_network to 0.27.6
This release contains a fix for fetching `NrsMapContainer` content.

There was a failure when this content was being retrieved because the
`set_sub_names` function was requiring that the 'label' for the URL be
63 characters or less. Since some recent changes, the `NrsMapContainer`
URLs are no longer less than 63 characters.

The function has now been changed to ignore the label validation.

The xor_name crate was upgraded to the same version as `safe_network`.

There were also changes in safe_network for configuration related to
qp2p, and the API was updated to handle these.