fix: Change default local address to `0.0.0.0`
As part of the recent refactoring of qp2p, the default local address was
changed to `127.0.0.1`, on the basis that it is more secure to require
users to opt-in to public reachability.

Sadly, this change was quite disruptive, and optimises the less common
case. Whilst it may still be desireable to default to local-only, the
specifics of how this is configured needs more thought.