feat: use new api for storing files
With the introduction of self encryption, we've now split the storage of files on the network out
into either the storage of a 'blob' or a 'spot'. A blob is anything that's over 3072 bytes and a
spot is anything less than that. Storage of spots don't use self encryption, whereas storage of
blobs do.