wip: upload build artifacts
The idea here is to have a separate `build` job, in order that the
caches get built even if tests fail spuriously. Sadly, this also
requires persisting the binaries between jobs – it's not enough to rely
on just the cache action for this, since a recovered cache will not be
overwritten (e.g. if only application code changes, not `Cargo.lock`).

This may actually slow down feedback, since it will block any tests
until all builds have finished.