chore: update `qp2p`
The new version moves the message sending methods from `Endpoint` on to
`Connection`, so we now need to obtain a `Connection` first via
`Endpoint::connect_to`.

For now, this still uses the connections pool, so the behaviour is
identical to before (e.g. get a connection from the pool or open and
store a new connection, then send a message). Further changes to `qp2p`
will start to remove the connection pool, meaning we will have to hold
on to relevant connections ourselves.