chore: upgrade sn_api to 0.35.x
The API is upgraded to 0.35.x. This API release still isn't necessarily a working release, and the
CLI tests are still failing at the point of this upgrade. We decided to merge the changes to the API
anyway, since there had been lots of changes outstanding for the past month.

Also fix quite a few clippy warnings and errors I had been meaning to address for a while.

BREAKING CHANGE: there have been commands that have been removed and we also can't upload things
like files with 0 bytes in them any more.