feat(bytes): encapsulate storage mechanism
`DataType` enum is simplified to `SafeKey` | `Bytes` | `Register`

Then `get_bytes` (taking `BytesAddress` enum) solves
which underlying storage mechanism is to be used
(`Spot`/`Blob`), as `BytesAddress` is either
`Spot(SpotAddress)` or
`Blob(BlobAddress)`

`upload_bytes` uses `bytes.len()` to choose underlying storage
mechanism, and returns `BytesAddress`.

`Bytes` is a simpler, more accurate and direct description
of the "data type", while `Data` is vague and ambiguous
(seems to be what all members of `DataType` are).
`Spot` and `Blob` are the two storage mechanisms for `Bytes`,
size dependent (due to optimisations) and simply meaning
- `As one chunk, optionally using encryption key`, or
- `As 4+ chunks, using self encryption`.

BREAKING CHANGE: Public enum `DataType` is modified, and APIs for
`Blob` and `Spot` are replaced with  corresponding for `Bytes`.