chore: update api for 0.33.x of safe_network
BREAKING CHANGE: Blob and Spot entries in the DataType enum are merged into a single Bytes entry.

The client API in the network has now been updated to encapsulate the Blob and Spot data. This makes
the client API much easier to use, as callers no longer need to be concerned about the size of the
data they're trying to store and retrieve.