feat: contact all known nodes and await AE-Retry if we only knew genesis
Cmds have been failing due to sending to only one node. This is due to
having no network knowledge, because we only contacted one peer in
bootstrap (which may fail).

So now we contact all known nodes if we have no prefixmap.

And if that is the case we'll await some knowledge before continuing.