chore: move sn_api code into an sn_api directory
The initial merge of the `sn_api` repository put the code for `sn_api` into the `src` directory at
the root of the repository, which has merged it in alongside the code for `safe_network`. We now
create an `sn_api` directory at the root of the repository and move the `sn_api` code in here. This
completely isolates the code for `sn_api` in its own crate, with the same Cargo.toml from the old
repository.

As things are, the crate builds fine.