chore: move safe_network code into sn directory
Due to the fact that we're now using multiple crates, the safe_network code is moved into the `sn`
directory.

A Cargo.toml is added to the root directory to establish this repository as a workspace, currently
with 2 members, sn and sn_api. If you now run a `cargo build` at the root directory, it will build
both of these crates.

The Github Actions workflows that were brought in from the `sn_api` merge were also removed here.