ci: update actions workflows for workspace refactor
Now that we have multiple crates in the same repository, if we only want to do things when subsets
of the source tree have been modified (e.g. only one crate particular crate has changed), it does
make the CI situation a bit more cumbersome. We need to make much more frequent use the path filter
action and only do things if certain files have been changed.

It's also useful for the workflows to indicate the crate they relate to, so they've been renamed or
added to accommodate that.

Certain things have been moved to the level of the workspace because with these, you can't really
discriminate based on the crate that's changed.

An unused dependency for sn_api was also taken out here.