ci: remove automated versioning workflows
We've agreed to move to using `cargo smart-release` for deployment and publishing for our new
workspace structure, so these workflows are being removed.

We might be able to call smart release in the bump version workflow in the future, but there might
be an issue regarding keeping the commit history clean, which may require human involvement. The
smart release tool fails on publishing sometimes, which causes multiple commit messages to be
generated. We're going to deal with this manually first, then possibly come back and re-enabled an
automated version bump and publish.

We definitely shouldn't need the other workflows, as the smart release tool does a merge and push to
master, so there's no need for opening new PRs.