refactor: update `qp2p`, which removes `ConnectionPool`
The latest release of `qp2p` has removed the `ConnectionPool`, meaning
connections now close immediately when both sides of it (`Connection`
and `ConnectionIncoming`). It's intended that we apply more deliberate
connection management as a substitute, but as a stop-gap this commit
implements a `ConnectedPeers` type, which behaves much like the `qp2p`
`ConnectionPool`.