fix: handle the case where we don't know an adult's capacity
This was left with a `todo!`, which has started triggering in tests.
Since there's only a single call-site, it's easy enough to return an
`Option` here and allow callers to make the decision on how to treat
missing info.

The one call-site we do have, is deciding how to handle non-specific
errors from adults when trying to retrieve a chunk. In this case we
treat the node as 'not full' in order to send the error back to the
client.

Separately, this seems likely to ignore other errors unrelated to node
capacity, so we might want a specific variant to cover this in future.