chore: move cli code into sn_cli crate directory
The first step after the initial merge was to reorganise some directories:

* The cli code was moved from the top level `src` to the `sn_cli` crate directory
* The test data and related directories were moved to the `resources` directory
* The test scripts were also moved to the `resources` directory

The CLI integration tests were updated to refer to the new test data locations. Most of the test
suite is now passing, but some NRS related tests weren't working, so these were marked as ignored.
These failures look to be due to some of the recent changes in the NRS refactor and will probably be
quite straight forward to fix. However, I'm going to address these after this merge.