fix: fix a possible panic when creating client
`client::connections::Session::make_contact_with_nodes` would slice the
known initial contacts, but this would panic if fewer than
`NODES_TO_CONTACT_PER_STARTUP_BATCH` (currently: 3) nodes are known.
It's not clear that this is a precondition or requirement, so for now we
simply take as many as we can up to
`NODES_TO_CONTACT_PER_STARTUP_BATCH`. If it *is* a requirement, we
should return an `Err` rather than panic.