refactor: update cli to use new nrs api
There are changes here to update the NRS command handlers to use the new API for NRS, and other
smaller changes for referencing types that had been moved to new modules. This commit also restores
a function, `get_map_summary`, that was deleted from the `NrsMap` struct. The CLI `cat` and `dog`
commands still relied on this functionality, so it's been provided as a placeholder for now, just to
get the CLI to compile.

At this point, the NRS test suite is not passing, because there have been some behavioural changes
as part of the refactor. Unfortunately, because of the state of the rest of the network just now,
it's hard to isolate these specific changes to get the tests passing again, as other functionality
like the `cat` command appear to be broken just now too.

I think it's worth committing this as it is now, with the CLI compiling, then we can come back and
address the other issues that will get the test suite passing again.