ci: configure basic pull request checks for sn_cli
Right now we just enable basic checks for `sn_cli` CI and delete the network test and the master
workflows. We will soon have a new release process which will incorporate much of what was in the
master workflow for sn_cli, but we can easily just pull the parts we need from the archived
`sn_cli` repo. The network tests workflow needed quite a bit of work anyway, and we can re-do/enable
that when we have a stable set of CLI tests.

The workflow for running a set of client tests has been renamed a little bit, just to make it more
clear that it relates to the `sn` crate.