ci(sn_api): remove `test-publish` job
`cargo publish --dry-run` depends on the dependencies versioned in
`Cargo.toml`. This is a problem in our new workspace world because the
version of `safe_network` that `sn_api` depends on will not be updated
until a version is actually published.

Non-versioned dependencies (e.g. path or git dependencies) cannot be
published to crates.io, so they cannot be tested with `cargo publish
--dry-run`. Therefore, this check has been removed from `sn_api`.