refactor: unit test node run-baby-fleming command
I wanted to extend the `run-baby-fleming` command to add a new network to the networks list, but
before doing so, I looked into providing test coverage to the node commands. Currently there are no
integration tests for these commands. The `node` commands don't really depend on having an existing
network, so they are much more susceptible to being unit tested.

The one barrier to unit testing was the fact that the command uses `sn_launch_tool` from the `sn`
crate, and this actually launches nodes. The solution was to introduce a `NetworkLauncher` trait and
provide a fake implementation of this trait for use in the unit tests. We use the fake launcher to
assert that the `sn_launch_tool` is called in the correct fashion. The concrete implementation for
use with the application was simply just a little wrapper around `sn_launch_tool`.

As part of this refactor, the `run-baby-fleming` command is changed as follows:
* The `node_path` argument was renamed to `node_dir_path`, to be clearer about the fact that this
  option is expected to be a directory path and not a file path.
* The `test` flag argument was removed, as we're not using the authentication service just now.
* The `verbose` flag argument was removed, as it only applied to the authentication service test,
  which was removed.

Due to the authentication service no longer being in use, the code related to performing a test
using the service was removed.

The unit tests provided here should give coverage for all the scenarios in `run-baby-fleming`.