feat: add baby fleming to networks list
The `node run-baby-fleming` command is updated to add the baby-fleming network to the networks list.

This is something that makes sense from a UX point of view, but it can also make it easier to use
some automation to add extra nodes to the network in an automated process that tries to get a local
network to settle to having 7 Elders, which is required to use some functionality for the network.
With the `baby-fleming` network added, it can be used as part of the `node join` command.