fix(sn): make `KvStore::store` perform atomic immutable writes
`KvStore` is the database abstraction we use to persist chunks. This is
itself backed by `sled`. `sled` uses a log as one of the underlying
storage primitives, which means that all inserts (even overwrites) will
consume additional space (until the log is compacted, which may happen
irregularly).

However, chunks themselves are write-once which means that we would only
ever need to write a value once. As such, we can use `compare_and_swap`
to write the value only if it's not currently set. Since chunks are
content-addressable, we furthermore don't need to do anything if the
value did already exist and wasn't written, since there is only one
possible value for a given key.