chore(sn): remove unused `KvStore::store_batch`
Although this was commented to indicate it will be used again soon, it
is trivial to recover code from git history, but it may be non-trivial
to maintain or fix unused code. That's the case for trying to fix used
space handling in `store_batch`, since `sled::Tree::apply_batch` cannot
perform compare-and-swap operations. As such, the whole operation
probably needs a rethink, and since it's not currently used we can defer
that until it's needed.