refactor(sn): add `KvStore::flush` to avoid waiting in tests
Some `KvStore` tests verify that used space adjusts appropriately when
data is written (also when data is removed, but those tests are disabled
since it's not predictable when sled will compact its log). We measure
used space by measuring the size of the database files.

sled flushes to disk asynchronously by default, and the tests were
accounting for this by looping until the size changed. This could lead
to hung tests in the case of bugs with used space calculation, which
would be harder to debug than a concrete failure. As such, a test-only
`flush` method has been added to `KvStore`, which is now used by the
tests to force a flush to disk.