feat: add a resources dir for `jq` filters
Using `jq` to process JSON logs could be a great tool for debugging
network activity. Particularly if we expand the use of fields with log
markers. This new directory is somewhere we can put useful filters for
reuse.

The first filter is one to filter for `ConnectionOpened` and
`ConnectionClosed` log markers, and output only the timestamp,
connection ID, and simplified spantrace for each.