refactor(sn)!: remove `reachable` from `messaging::system::Peer`
Although the field is `false` on initialisation, and it is checked in a
couple of places, in every case as far as I can tell it is set to `true`
right after construction. As such, it's not really doing anything for us
and we can remove it.

BREAKING CHANGE: The `reachable` field has been removed from
`messaging::system::Peer`. The `set_reachable` and `is_reachable`
methods have been removed from `routing::PeerUtils`. Note that this
affects the message format, since `Peer` is included in `NodeState`,
which appears in various messages.