refactor(sn)!: remove `Peer` from `NodeState`
Syntactially, having a "peer" struct inside `NodeState` doesn't make a
lot of sense, vs. including the name and address directly in the state.
This also unblocks moving `Peer` out of `messaging`, which will remove
the requirement for it to be serialisable etc.

To make migration easier, an `age` method was added to the
`NodeStateUtils` trait.

BREAKING CHANGE: The `peer` field has been removed from `NodeState`, and
has been replaced by `name` and `addr` fields. This also affects the
message format, since `NodeState` appears in several messages.