refactor(sn)!: remove `PeerUtils` trait
This trait has been redundant since the `sn_messaging` and `sn_routing`
crate were merged, but it's doubly so when the `Peer` struct is now
itself in `routing`.

BREAKING CHANGE: The `routing::PeerUtils` trait has been removed. The
methods it provided are now available as an inherent impl on
`routing::Peer`.