refactor(sn): replace `(XorName, SocketAddr)` with `Peer`
In many places, we pass around `Peer`s which were then converted to
`(XorName, SocketAddr)`. Most of the time this ends up at the
`Comm::send*` methods. These methods have been updated to take `Peer`,
and this change has been propagated to anything else that was storing or
passing `(XorName, SocketAddr)`. Overall this is a nice simplification,
and removes more conversions than it introduces.