chore(sn_api): revert switch from `time` to `chrono`
Revert "chore(sn_api): switch from `chrono` to `time`".

This reverts commit 442c1bdd3f9c22424aa9642d71d4016c868b0b58.

This change breaks publishing for the `sn_api` crate, which seems to insist on bringing in some
other version of the crate.

The change was made because RUSTSEC-2020-0159 (security advisory item) was being flagged by
`cargo-deny`. It seems to be quite acceptable to ignore this issue, so it can be added to an ignore
list.