refactor(sn)!: remove `Copy` from `Peer`
Eventually we may want to put a `Connection` into `Peer`, which is not
`Copy`. As such, `Peer` itself will be unable to derive `Copy`. Although
this might not happen for a while, removing dependence on `Copy` now
will make it smoother when the time comes.

BREAKING CHANGE: `routing::Peer` is no longer `Copy`.