refactor(sn): pass `Connection` to `join_network`
When bootstrapping to a network, we immediately connect to a peer in
`Comm::bootstrap`. Before this commit we would essentially throw this
away. We now return it, and pass it on into `join_network` so that the
initial join request reuses the connection.

To make the APIs work nicely, and support more connection reuse in
future, `Peer` now has a `connection: Option<Connection>` field, and
this is used by `Comm::send` to send over the existing connection if one
is set.

For a two node network, this only really saves a single connection,
since the first join request will get a retry response, and we lose the
connection handle when this happens. This may be improved in future, if
the gains seem worthwhile. Another caveat is that `Peer::connection` may
be `Some`, but the contained connection may be lose. This is not
currently handled by `Comm::send` (the send will simply fail).