refactor(sn): support reusing connections from incoming messages
This replaces the `SocketAddr` in `ConnectionEvent` with a new
`UnknownPeer` struct, which always has an address and may have a
`Connection` (currently, the `Connection` would only be `None` in tests,
but that might not always be the case). An `UnknownPeer` can be
'upgraded' to `Peer` by `identify`ing it with a name.

This is currently only used in the join flow.