ci: release workflow for safe_network crate
The release process is largely the same, but with the introduction of the workspace I've just taken
the opportunity to tidy up the workflow a little bit.

To make it a bit more descriptive, the workflow was renamed to 'release' rather than 'main'.

This workflow will be extended to include the changelogs and assets for `sn_api` and `sn_cli`.