ci: release workflow for sn_api crate
The release workflow has been extended to include the sn_api crate.

When there's a release, it's possible that any one, or many crates could have new releases, and the
conditions used on jobs have been updated to take this into account. There's been additional
scripting added for the changelogs and the release and tag names to be used, that also take this
into account.