refactor(sn): retry on connection loss of reused connection
The `routing::Comm::send` method uses an instance of `ConnectedPeers` to
reuse existing connections. This saves a lot of connections during busy
exchanges between few nodes, such as DKG. However, should the connection
retrieved for reuse have become invalid (e.g. closed by the peer or
timed out), that recipient would be recorded as failed.

Since connection timeouts are to be expected, particularly for stored
connections with keep-alives disabled as in `ConnectedPeers`, it makes
sense to retry the recipient with a fresh connection before declaring
it failed.

The implementation has been bolted on to the existing `send` mechanism.
The `send` closure will return an additional boolean indicating whether
or not the connection was reused. When evaluating the result, if it
failed due to connection loss on a reused connection, we crate a new
`send` task for the same recipient, but pass an additional flag to force
a new connection. Thus, if the peer fails *again* for connection loss,
the connection will not have been reused and the peer will be recorded
as a failed recipient as normal.