ci: use smart-release to auto bump version
We change from using our own custom action that used the `standard-version` Node library to perform
semantic version bumps and generate changelogs, to `cargo smart-release`.

The `smart-release` tool uses the presence of 'conventional commits' to:
* Identify crates in the workspace that have changes
* Perform semantic version bumps
* Update any dependent crates to use bumped version numbers
* Generate changelogs

The changes it makes are captured in a commit it generates, and for each crate that had a change, it
also creates tags at this commit. We soft reset this commit so we can change the commit message to
use `chore(release): `, re-commit, then update the tags to point to this new commit. After this, we
push both the release commit and the tags.

Previously we had separate workflows for the version bumping and tagging. The version bump commit
was submitted in an automated PR, which was then merged, and kicked off the tagging workflow, which
tagged and pushed the tags. We now don't have the generated PR and just push the changes directly.