refactor(sn): add an `Option<qp2p::Connection>` to `Peer`
This allows the `Peer` struct to carry an existing connection to the
peer, if one exists. Currently, the connection must be set on
construction with `Peer::connected`. A reference to the connection can
be retrieved with `Peer::connection`. When sending a message using
`Comm::send` or `Comm::send_on_existing_connection`, we now check the
`Peer` before `ConnectedPeers`.

We never actually set `connection` currently, but there are several
possible next steps that would all depend on `Peer` carrying a
`connection`, so introducing it separately makes sense.