refactor(sn): represent `HandleMessage::sender` as an enum
This is a precursor to having `HandleMessage::sender` contain the
originating `Connection` for connected peers. Since we use
`HandleMessage` also when 'sending' messages to ourselves, we have to
deal with the possibility that there may not be a connection. We could
simply use `Option` for this, however the dedicated enum makes it
clearer exactly what situation is represented.

To facilitate later plumbing, `ConnectionEvent` now also uses `Sender`
to represent the message sender.

It's possible that we should make more use of the knowledge that a
message was sent to ourselves, e.g. to short-circuit AE checks. For now,
we more-or-less immediately lookup our own address, and proceed as if
the message was actually sent from there.