refactor(sn): correlate client connections without `ConnectedPeers`
This is a first pass at introducing a mechanism to correlate client
connections without using the `ConnectedPeers` cache/pool. The approach
is centred on passing along the connected `Peer` that originated the
request, which in most cases is enough to get the response to reply
directly to the same `Peer` instance.

Chunk queries need an additional mechanism, since the elder cannot reply
right away, and instead must forward the query to nodes, and can only
reply to the client once the nodes have themselves replied. For this we
introduce a `pending_chunk_queries` field to `routing::Core`. When a
chunk query is received from a client, a random `XorName` is generated
as a correlation ID, and the client `Peer` is stored in the
`pending_chunk_queries` cache under that `XorName`. The `XorName` is
then passed as the 'end user' for the message to nodes, who send it back
in their reply, allowing the receiving elder to correlate the response
with the origin client `Peer`, and forward the response to the client
over the `Peer`'s connection.

There are some minor behavioural changes with this approach:

1. Unlike the connection pooling approach, the client's exact connection
   must remain valid (e.g. if they reconnect while waiting for a
   response, they will never receive it). There is no way around this
   without a general-purpose connection cache, since we do not know on
   connection whether a peer has ever sent a client request. Thankfully
   it's not a big deal, since client operations are always idempotent,
   so client logic can handle any connection interruption by sending the
   same request(s) over a new connection.

2. The `pending_chunk_queries` cache uses duration-based expiration,
   currently hard-coded to 5 minutes. Previously there was no timeout
   for this operation, and it would just depend on the connection
   remaining open in the pool. Since we cannot easily detect connection
   closure in `routing::Core`, and 5 mins is longer than the default
   client query timeout, this seems reasonable for now.