ci: prevent release workflow from running
The release workflow now requires the release commit before it will run. This is to prevent the
build stage from running on any push to main, which will include a PR merge that doesn't have a
release commit.

Also added some debugging text for the version bump workflow, which is still failing on the GHA
build agent.