refactor(sn): set `Peer::connection` on send
This changes the `Peer::connection` field to an `Arc<RwLock<...>>` so
that `routing::Comm` can set the connection if one is opened. This
shaves off ~ 4k `ConnectionOpened` log markers when forming an 11 node
network.

The main concern with this approach is that any existing connection is
simply overwrote, and so will be disassociated from the `Peer`. This
should currently be fine, since we do not rely on messages coming from
particular connections.