refactor(sn): store a `Peer` in `NodeState`
Since `Peer` can now hold a `connection`, we'd like to get `Peer`
instances into as many places as possible to enable connection reuse.
`NodeState` records are held by `SectionPeers`, which contains all known
members of a node's section. If we can get a connection in there, we
should be able to reuse it for most node<->node communication.

The change was largely mechanical. The only interesting thing is that
`serde::Serialize` has to be implemented manually, since `Peer` is not
serializable. The implementation uses `messaging::NodeState`
internally, to ensure consistency with that format (which is necessary
for a `SectionAuth<NodeState>` to verify).