refactor(sn): reuse known `Peer`s in DKG sessions
Now that `NodeState` contains a `Peer`, which itself may contain an
existing `Connection`, it makes sense to try and reuse the `Peer`s in
network knowledge when initialising DKG sessions.

This doesn't have much impact on the total number of connections yet,
probably because we don't always maintain the `Peer`s in network
knowledge.