refactor(sn): store `Peer`s in `SectionAuthorityProvider`
`SectionAuthorityProvider` contains details of the elders of a section,
and is often used as the source of recipients when sending messages to
elders. By storing a `Peer` inside the SAP, we ensure that connections
will be reused if possible when messaging elders.

Note that we wholesale replace the SAP in a few cases, which will
immediately cause us to drop connections. In future we can be more
careful to retain or selectively update existing `Peer`s to get more
reuse.