ci: check each crate for changes individually
We need to determine if each individual crate has changes and then only run smart-release with the
crates that do have changes. The way it behaves varies based on whether there's changes to just one
crate, or many, and which crates you specify in the command.

It's a shame you can't just run it without a crate name and have the tool determine which of the
crates have changes. I wonder if this is due to the fact that the safe_network crate doesn't
correspond to the 'sn' directory name, but not sure.