refactor(sn): replace `Sender` with `UnknownPeer`
The `UnknownPeer` struct is more general, and will make it easier to
propagate connections into more places (e.g. this already propagates
connections into `HandleSystemMsg`). It's also generally easier to use
than `Sender`, and doesn't need any special casing for test-only
purposes.

The downside is that we lose the 'sending to ourself' representation.
This wasn't being used so far though, so it's only a hypothetical. The
comment on `UnknownPeer` attempts to document that 'rough edge' (since
we would otherwise expect to only know peers either through
advertisement, where the name would be known, or through direct
connection, where the connection would be present).