chore(sn): add a feature for increased `WireMsg` debug info
When we serialize to a `WireMsg`, we lose visibility of the exact
contents of the message. Serialization is necessary in order to obtain a
signature for the message, so we can't simply hold the message
unserialised until send time (at least, not without passing around a
keypair, which we don't want to do).

As a hacky alternative, an `unstable-wiremsg-debuginfo` feature has been
added. When enabled, this adds a `payload_debug` field to `WireMsg`
which can be set to contain a debug representation of the message
contents. Since `WireMsg` is constructed with `Bytes`, its necessary for
callers to set this themselves.