refactor(sn): duplicate `Proposal` in `routing`
This simplifies a bunch of conversion to/from the `messaging` and
`routing` variants of `NodeState` and `SectionAuthorityProvider`.
Additionally, this lets us reuse more connections when we handle a
proposal ourselves, such as when responding to a `JoinRequest`.

This had a surprisingly high impact on connection use, shaving off 500
or so when forming an 11 node network.