ci: use different tag name for gh release
In the last attempt to perform a release, the creation of the Github Release failed. Unfortunately
there was no error output from the action, so I'm left guessing as to the cause. I suspect it might
be because the tag name being used has a '/' in it. However, I'm not completely sure about this,
because I created a release on my fork using the gh command line tool, and I used the same tag name
as the run that had the error. The release creation was successful, but I notice it did return back
a strange URL, so this is my best bet at the moment.